+++ flutter_readium/android/src/main/kotlin/dk/nota/flutterreadium/ReadiumReader.kt
@@ -661,10 +661,6 @@ object ReadiumReader :
     }
 
     fun closePublication() {
-        _currentPublication?.close()
-        _currentPublication = null
-        currentPublicationCssSelectorMap = null
-
         ttsNavigator?.dispose()
         ttsNavigator = null
         audiobookNavigator?.dispose()
@@ -677,6 +673,10 @@ object ReadiumReader :
 
         currentReadiumTimebasedState.value = ReadiumTimebasedState()
         currentTextLocator.value = null
+
+        _currentPublication?.close()
+        _currentPublication = null
+        currentPublicationCssSelectorMap = null
 
         state.clear()
     }
@@ -1454,11 +1454,7 @@ object ReadiumReader :
     ) {
         val navigator = epubNavigator ?: return
         withMainContext {
-            if (navigator.preferences?.disableSynchronization == true) {
-                return@withMainContext
-            }
-
-            navigator.goToLocator(locator, animated, segmentDuration)
+            navigator.syncToLocator(locator, animated, segmentDuration)
         }
     }